Properly kill containerized linters
We can't simply SIGKILL a job that's containerized -- especially
as we'd just be terminating the docker *client*, not the linter process
itself.  (SIGTERM and the like may be proxied, but by its nature SIGKILL
cannot be.)  This commit updates `ale#job#Stop()` to have docker look up
the container id by label, then perform the killing via `docker kill`.

This is also a fairly rough commit.  It's working, but there are a
number of things to go back and address (e.g. neovim support).